+++ frontend/src/translate/languages/pt.js
@@ -720,7 +720,8 @@ const messages = {
         }
       },
       notifications: {
-        noTickets: "Nenhuma notificação."
+        noTickets: "Nenhuma notificação.",
+        dismiss: "Dispensar notificação"
       },
       quickMessages: {
         title: "Respostas Rápidas",